moved code and new test
