+++ eq-author/src/App/QuestionnaireDesignPage/MainNavigation/index.test.js
@@ -79,7 +79,7 @@ describe("MainNavigation", () => {
   });
 
   it("should disable qcodes, publish and preview buttons if there are errors on questionnaire", () => {
-    const { getByTestId } = defaultSetup({ changes: { totalErrorCount:2 } });
+    const { getByTestId } = defaultSetup({ changes: { totalErrorCount: 2 } });
     expect(getByTestId("main-navigation")).toBeTruthy();
 
     expect(getByTestId("btn-preview").hasAttribute("disabled")).toBeTruthy();
@@ -91,12 +91,12 @@ describe("MainNavigation", () => {
   });
 
   it("should NOT disable qcodes button if there is no surveyId", () => {
-    const { getByTestId } = defaultSetup({ changes: { totalErrorCount:1 } });
+    const { getByTestId } = defaultSetup({ changes: { totalErrorCount: 1 } });
     expect(getByTestId("btn-qcodes").hasAttribute("disabled")).toBeFalsy();
   });
 
   it("should disable qcodes button if there is no surveyId BUT there is another error", () => {
-    const { getByTestId } = defaultSetup({ changes: { totalErrorCount:2 } });
+    const { getByTestId } = defaultSetup({ changes: { totalErrorCount: 2 } });
     expect(getByTestId("btn-qcodes").hasAttribute("disabled")).toBeTruthy();
   });
