+++ eq-author/src/App/QuestionnaireDesignPage/index.test.js
@@ -1,10 +1,11 @@
 import React from "react";
-import { render } from "tests/utils/rtl";
+import { render, userEvent, fireEvent } from "tests/utils/rtl";
 import { buildQuestionnaire } from "tests/utils/createMockQuestionnaire";
 import { useQuery } from "@apollo/react-hooks";
 import { MeContext } from "App/MeContext";
 import { PAGE } from "constants/entities";
 import { useParams } from "react-router-dom";
+import hotkeys from "hotkeys-js";
 
 import QuestionnaireDesignPage from "./";
 import useQuestionnaireQuery from "./useQuestionnaireQuery";
@@ -159,6 +160,25 @@ describe("QuestionnaireDesignPage", () => {
     it("should display questionnaire title if no longer loading", () => {
       setup();
       expect(document.title).toEqual("questionnaire - Page 1.1.1");
+    });
+  });
+
+  describe("SuperNav", () => {
+    it("should focus on a nav block when the hotkey F6 button is pressed and move forward", () => {
+      const { getByTestId } = setup();
+      const focusSuperNav = "SuperNav-1";
+      // const domNode = document.getElementById("SuperNav-3");
+
+      fireEvent.keyPress(document.body, {
+        key: "F6",
+        code: "F6",
+        keyCode: 117,
+        charCode: 117,
+      });
+
+      // console.log(defaultSetup());
+      expect(document.getElementById(focusSuperNav)).toHaveFocus();
+      // expect(getByTestId("SuperNav")).toHaveFocus();
     });
   });
 });
